Separate MediaPlayer components out a bit further (to simplify eventual migration to Psammead)
