+++ src/app/containers/MediaPlayer/helpers/placeholder.js
@@ -1,0 +1,18 @@
+const getPlaceholderSrc = src => {
+  const parts = src.split('/');
+  const [domain, media, imgService, width, ...extraParts] = parts;
+  const definedWidth = width.replace('$width', '512');
+  const domainWithProtocol = `https://${domain}`;
+
+  const newUrl = [
+    domainWithProtocol,
+    media,
+    imgService,
+    definedWidth,
+    ...extraParts,
+  ];
+
+  return newUrl.join('/');
+};
+
+export default getPlaceholderSrc;
