+++ src/app/containers/MediaPlayer/index.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import pathOr from 'ramda/src/pathOr';
-import styled from 'styled-components';
 import AmpMediaPlayer from './amp';
 import CanonicalMediaPlayer from './canonical';
-import MediaPlayerPlaceholder from './placeholder';
+import MediaPlayerWrapper from './wrapper';
 import Metadata from './Metadata';
 import embedUrl from './helpers/embedUrl';
+import getPlaceholderSrc from './helpers/placeholder';
 import filterForBlockType from '../../lib/utilities/blockHandlers';
 import useToggle from '../Toggle/useToggle';
 import { RequestContext } from '../../contexts/RequestContext';
@@ -14,32 +14,6 @@ import {
   mediaPlayerPropTypes,
   emptyBlockArrayDefaultProps,
 } from '../../models/propTypes';
-
-const landscapeRatio = '56.25%'; // (9/16)*100 = 16:9
-const portraitRatio = '177.78%'; // (16/9)*100 = 9:16
-const StyledContainer = styled.div`
-  padding-top: ${({ orientation }) =>
-    orientation === 'Portrait' ? portraitRatio : landscapeRatio};
-  position: relative;
-  overflow: hidden;
-`;
-
-const placeholderImage = src => {
-  const parts = src.split('/');
-  const [domain, media, imgService, width, ...extraParts] = parts;
-  const definedWidth = width.replace('$width', '512');
-  const domainWithProtocol = `https://${domain}`;
-
-  const newUrl = [
-    domainWithProtocol,
-    media,
-    imgService,
-    definedWidth,
-    ...extraParts,
-  ];
-
-  return newUrl.join('/');
-};
 
 const MediaPlayerContainer = ({ blocks, placeholder }) => {
   const { id, platform, origin } = React.useContext(RequestContext);
@@ -72,8 +46,8 @@ const MediaPlayerContainer = ({ blocks, placeholder }) => {
   }
 
   const Player = isAmp ? AmpMediaPlayer : CanonicalMediaPlayer;
-  const placeholderSrc = placeholderImage(imageUrl);
   const shouldShowPlaceholder = !isAmp && placeholder;
+  const placeholderSrc = getPlaceholderSrc(imageUrl);
   const embedSource = embedUrl({
     vpid: versionId,
     assetId: id,
@@ -84,15 +58,12 @@ const MediaPlayerContainer = ({ blocks, placeholder }) => {
   return (
     <GridItemConstrainedMedium>
       <Metadata aresMediaBlock={aresMediaBlock} />
-      <StyledContainer>
-        {shouldShowPlaceholder ? (
-          <MediaPlayerPlaceholder src={placeholderSrc}>
-            <Player embedSrc={embedSource} />
-          </MediaPlayerPlaceholder>
-        ) : (
-          <Player embedSrc={embedSource} />
-        )}
-      </StyledContainer>
+      <MediaPlayerWrapper
+        showPlaceholder={shouldShowPlaceholder}
+        src={placeholderSrc}
+      >
+        <Player embedSrc={embedSource} />
+      </MediaPlayerWrapper>
     </GridItemConstrainedMedium>
   );
 };
